backend/requirements.txt
alembic>=1.16,<2.0
fastapi>=0.116,<1.0
httpx>=0.28,<1.0
pandas>=2.2,<3.0
pydantic-settings>=2.10,<3.0
psycopg[binary]>=3.2,<4.0
pytest>=8.4,<9.0
python-dotenv>=1.1,<2.0
sqlalchemy>=2.0.41,<3.0
uvicorn[standard]>=0.35,<1.0
